merge TSSC and TISC into same repo refactor (#13)
Everything looks fine. There are some things that I want to swap around, but we can cover that in another feature. 

Once I'm familiar with some of the inheritance - we'll likely want to have a "common"/"overall" section + the individual components.

Additionally, we might want to put folders for each group under the "docs" section as we compound more "tssc" "cnad" and "tisc" content (./docs will fill up quick). But that's a bridge we can cross once we get there.